StatisticsActivityTest.kt 추가
WeekSummaryWorkerTest.kt 수정 (coverage 향상)

+++ Sumdays/app/src/test/java/com/example/sumdays/statistics/TestParcelableData.kt
@@ -18,7 +18,7 @@ data class SummaryDetails(
 data class EmotionAnalysis(
     val distribution: Map<String, Int>,
     val dominantEmoji: String,
-    val emotionScore: Float,
+    val emotionScore: Double,
     val trend: String? = null
 ) : Parcelable
